Create better family, primary adult, and user creation workflow
Specifically, this commit splits the family _form.html.erb file
into two separate files, _edit.form.html.erb and _new.form.html.erb.
This enables us to have different logic for creating and editing
families that we'd need to use javascript for if only using one
create/edit view.

Also, we try to embed the creation of the first person in the family
into the family#new form so that when a user creates their family,
they also create themselves as a person in that family.

This is still a work in progress, so more work needs to be done to
make sure everything is working correctly and create passing tests.