Refactor family/person creation
This builds on the goal of having the primary adult or person
attending CUCO added automatically when one creates their family.
This functionality is now working, however it's no longer possible
to edit one's newly created family.

Also, this commit starts the removal of the "This is me" workflow
by first removing the checkbox from the person#edit form. More
work on this needs to be done so that the logic is removed from
the people controller as needed.